Functional GUI controller
This programme has functional GUI controller which is integrated with Model and View. GUI view have been altered to show all threads on GUIMain and Report (flight capacity report) on GUI Report. GUIMain has a 'Open Checkin' button which when pressed executes the threads. The readout though is still visible on the console.
Pending- to display the readout of threads execution on the GUIMain respective panels.