Something visible on GUI view + functional GUI controller
This version has something visible in GUI view and as the previous commit has a functional GUI controller. The GUIMain 'Open Check In' button executes the threads.
Next challenge is to divert the thread based text printed on the console to GUIMain text areas. May have to use Swingworker class to publish() threads onto GUIMain as they progress. 
Overall there are three ways to do this and it is likely that swing worker would be ideal way to do this (will update more when I know more about this:
1. get the string report from the void method (write a toString method in PassengersInQueue class
2. divert the console text to GUI text area by using a new string. GUIMain has this code but is not tested fully for functionality (eg. need to declare it as an independent thread- please read the comments in the code)
3. use Swingworker class to publish() thread output as and when it happens 
Let me know what you think. I will continue to work on. However, it will be good if we work on this in an agile manner.