Prints one name
So basically we were on the right track. The issue lies in that none of the methods in the thread return any of the variables. Once they do this, we can run them in a swingworker thread and update the GUI using the publish() method, which can then be accessed using the process() method.